[HAL] Move AMD executable targets out of drivers
AMDGPU target IDs and executable compatibility are shared protocol
vocabulary used by both native drivers and HAL clients. Keeping that
code under the AMDGPU driver created a forbidden HIP-to-AMDGPU
dependency and made clients consume a native driver utility.

Move the parser, compatibility rules, generated map, and selector into
a neutral HAL executable package. Drivers and clients now depend
downward on that package, whose closure contains no HAL driver code.
