+++ build_tools/amdgpu/target_map.py
@@ -549,7 +549,7 @@ def render_cmake():
 
 
 def render_target_id_inl(repo_root):
-    output_path = "runtime/src/iree/hal/drivers/amdgpu/util/target_id_map.inl"
+    output_path = "runtime/src/iree/hal/executable/amdgpu/target_id_map.inl"
     lines = [
         generated_header("//", output_path),
         "//",
@@ -663,13 +663,13 @@ def render_header():
 
 def generated_outputs(repo_root):
     build_tools_output_dir = repo_root / "build_tools/amdgpu"
-    util_output_dir = repo_root / "runtime/src/iree/hal/drivers/amdgpu/util"
+    target_output_dir = repo_root / "runtime/src/iree/hal/executable/amdgpu"
     return {
         build_tools_output_dir / "target_map.bzl": render_bzl(),
         build_tools_output_dir / "target_map.cmake": render_cmake(),
         build_tools_output_dir / "elf_machine_map.inl": render_elf_machine_map_inl(),
         build_tools_output_dir / "target_map.h": render_header(),
-        util_output_dir / "target_id_map.inl": render_target_id_inl(repo_root),
+        target_output_dir / "target_id_map.inl": render_target_id_inl(repo_root),
     }
