+++ build_tools/lefthook/presubmit.py
@@ -1189,7 +1189,7 @@ def run_amdgpu_target_map(paths: list[str], fix: bool, verbose: bool) -> bool:
     relevant_prefixes = (
         "build_tools/amdgpu/elf_machine_map.inl",
         "build_tools/amdgpu/target_map.",
-        "runtime/src/iree/hal/drivers/amdgpu/util/target_id_map.inl",
+        "runtime/src/iree/hal/executable/amdgpu/target_id_map.inl",
     )
     if not any(path.startswith(relevant_prefixes) for path in paths):
         return skip_step("AMDGPU target map", "no AMDGPU target-map inputs")
@@ -1204,7 +1204,7 @@ def run_amdgpu_target_map(paths: list[str], fix: bool, verbose: bool) -> bool:
                 "build_tools/amdgpu/target_map.bzl",
                 "build_tools/amdgpu/target_map.cmake",
                 "build_tools/amdgpu/target_map.h",
-                "runtime/src/iree/hal/drivers/amdgpu/util/target_id_map.inl",
+                "runtime/src/iree/hal/executable/amdgpu/target_id_map.inl",
             ],
             verbose,
         )
