+++ runtime/src/iree/hal/drivers/amdgpu/BUILD.bazel
@@ -28,8 +28,8 @@ iree_runtime_cc_library(
     deps = [
         "//runtime/src/iree/base",
         "//runtime/src/iree/hal",
-        "//runtime/src/iree/hal/drivers/amdgpu/util:executable_target",
-        "//runtime/src/iree/hal/drivers/amdgpu/util:target_id",
+        "//runtime/src/iree/hal/executable/amdgpu:executable_target",
+        "//runtime/src/iree/hal/executable/amdgpu:target_id",
         "//runtime/src/iree/hal/utils:device_spec_builder",
     ],
 )
@@ -293,9 +293,9 @@ iree_runtime_cc_library(
         "//runtime/src/iree/hal/drivers/amdgpu/util:pm4_program",
         "//runtime/src/iree/hal/drivers/amdgpu/util:queue_primitives",
         "//runtime/src/iree/hal/drivers/amdgpu/util:shadow_map",
-        "//runtime/src/iree/hal/drivers/amdgpu/util:target_id",
         "//runtime/src/iree/hal/drivers/amdgpu/util:topology",
         "//runtime/src/iree/hal/drivers/amdgpu/util:vmem",
+        "//runtime/src/iree/hal/executable/amdgpu:target_id",
         "//runtime/src/iree/hal/memory:passthrough_pool",
         "//runtime/src/iree/hal/memory:slab_provider",
         "//runtime/src/iree/hal/memory:tlsf_pool",
@@ -383,9 +383,9 @@ iree_runtime_cc_library(
         "//runtime/src/iree/hal/drivers/amdgpu/util:pm4_program",
         "//runtime/src/iree/hal/drivers/amdgpu/util:queue_primitives",
         "//runtime/src/iree/hal/drivers/amdgpu/util:shadow_map",
-        "//runtime/src/iree/hal/drivers/amdgpu/util:target_id",
         "//runtime/src/iree/hal/drivers/amdgpu/util:topology",
         "//runtime/src/iree/hal/drivers/amdgpu/util:vmem",
+        "//runtime/src/iree/hal/executable/amdgpu:target_id",
         "//runtime/src/iree/hal/memory:slab_provider",
         "//runtime/src/iree/hal/memory:tracing",
     ],
@@ -600,8 +600,8 @@ iree_runtime_cc_test(
         "//runtime/src/iree/hal/cts/util:test_base",
         "//runtime/src/iree/hal/drivers/amdgpu/cts:testdata_amdgpu",
         "//runtime/src/iree/hal/drivers/amdgpu/cts:tsan_executable_testdata_amdgpu",
-        "//runtime/src/iree/hal/drivers/amdgpu/util:executable_target",
         "//runtime/src/iree/hal/drivers/amdgpu/util:packet",
+        "//runtime/src/iree/hal/executable/amdgpu:executable_target",
         "//runtime/src/iree/io:file_handle",
         "//runtime/src/iree/testing:gtest",
         "//runtime/src/iree/testing:gtest_main",
@@ -625,8 +625,8 @@ iree_runtime_cc_test(
         "//runtime/src/iree/hal/cts/util:registry",
         "//runtime/src/iree/hal/cts/util:test_base",
         "//runtime/src/iree/hal/drivers/amdgpu/cts:testdata_amdgpu",
-        "//runtime/src/iree/hal/drivers/amdgpu/util:executable_target",
         "//runtime/src/iree/hal/drivers/amdgpu/util:packet",
+        "//runtime/src/iree/hal/executable/amdgpu:executable_target",
         "//runtime/src/iree/testing:gtest",
         "//runtime/src/iree/testing:gtest_main",
     ],
@@ -670,7 +670,7 @@ iree_runtime_cc_benchmark(
         "//runtime/src/iree/hal/drivers/amdgpu/abi",
         "//runtime/src/iree/hal/drivers/amdgpu/util:benchmark_flags",
         "//runtime/src/iree/hal/drivers/amdgpu/util:benchmark_profile",
-        "//runtime/src/iree/hal/drivers/amdgpu/util:executable_target",
+        "//runtime/src/iree/hal/executable/amdgpu:executable_target",
     ],
 )
