+++ runtime/src/iree/hal/drivers/amdgpu/cts/BUILD.bazel
@@ -40,7 +40,7 @@ iree_runtime_cc_library(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/cts/util:registry",
         "//runtime/src/iree/hal/drivers/amdgpu",
-        "//runtime/src/iree/hal/drivers/amdgpu/util:executable_target",
+        "//runtime/src/iree/hal/executable/amdgpu:executable_target",
     ],
 )
