+++ runtime/src/iree/hal/drivers/amdgpu/profile_counters.h
@@ -11,7 +11,7 @@
 #include "iree/hal/drivers/amdgpu/util/aql_emitter.h"
 #include "iree/hal/drivers/amdgpu/util/aql_ring.h"
 #include "iree/hal/drivers/amdgpu/util/libaqlprofile.h"
-#include "iree/hal/drivers/amdgpu/util/target_id.h"
+#include "iree/hal/executable/amdgpu/target_id.h"
 #include "iree/hal/profile_schema.h"
 #include "iree/hal/profile_sink.h"
