+++ runtime/src/iree/hal/drivers/amdgpu/util/BUILD.bazel
@@ -229,45 +229,14 @@ iree_runtime_cc_library(
 )
 
 iree_runtime_cc_library(
-    name = "target_id",
-    srcs = ["target_id.c"],
-    hdrs = ["target_id.h"],
-    textual_hdrs = ["target_id_map.inl"],
-    deps = [
-        "//runtime/src/iree/base",
-    ],
-)
-
-iree_runtime_cc_library(
-    name = "executable_target",
-    srcs = ["executable_target.c"],
-    hdrs = ["executable_target.h"],
-    deps = [
-        ":target_id",
-        "//runtime/src/iree/hal",
-        "//runtime/src/iree/hal/utils:device_spec_builder",
-    ],
-)
-
-iree_runtime_cc_test(
-    name = "executable_target_test",
-    srcs = ["executable_target_test.cc"],
-    deps = [
-        ":executable_target",
-        "//runtime/src/iree/testing:gtest",
-        "//runtime/src/iree/testing:gtest_main",
-    ],
-)
-
-iree_runtime_cc_library(
     name = "code_object_target",
     srcs = ["code_object_target.c"],
     hdrs = ["code_object_target.h"],
     textual_hdrs = ["//build_tools/amdgpu:elf_machine_map.inl"],
     deps = [
-        ":target_id",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base:core_headers",
+        "//runtime/src/iree/hal/executable/amdgpu:target_id",
         "//runtime/src/iree/hal/utils:elf_format",
     ],
 )
@@ -277,9 +246,9 @@ iree_runtime_cc_test(
     srcs = ["code_object_target_test.cc"],
     deps = [
         ":code_object_target",
-        ":target_id",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base:core_headers",
+        "//runtime/src/iree/hal/executable/amdgpu:target_id",
         "//runtime/src/iree/testing:gtest",
         "//runtime/src/iree/testing:gtest_main",
     ],
@@ -321,9 +290,9 @@ iree_runtime_cc_library(
     hdrs = ["pm4_dispatch.h"],
     deps = [
         ":packet",
-        ":target_id",
         "//runtime/src/iree/base",
         "//runtime/src/iree/hal/drivers/amdgpu/abi",
+        "//runtime/src/iree/hal/executable/amdgpu:target_id",
     ],
 )
 
@@ -439,8 +408,8 @@ iree_runtime_cc_library(
     srcs = ["device_library_target.c"],
     hdrs = ["device_library_target.h"],
     deps = [
-        ":target_id",
         "//runtime/src/iree/base",
+        "//runtime/src/iree/hal/executable/amdgpu:target_id",
     ],
 )
 
@@ -460,8 +429,8 @@ iree_runtime_cc_library(
         ":pm4_dispatch",
         ":pm4_program",
         ":queue_primitives",
-        ":target_id",
         ":vmem",
+        "//runtime/src/iree/hal/executable/amdgpu:target_id",
     ],
 )
 
@@ -538,17 +507,6 @@ iree_runtime_cc_test(
     srcs = ["device_library_target_test.cc"],
     deps = [
         ":device_library_target",
-        "//runtime/src/iree/base",
-        "//runtime/src/iree/testing:gtest",
-        "//runtime/src/iree/testing:gtest_main",
-    ],
-)
-
-iree_runtime_cc_test(
-    name = "target_id_test",
-    srcs = ["target_id_test.cc"],
-    deps = [
-        ":target_id",
         "//runtime/src/iree/base",
         "//runtime/src/iree/testing:gtest",
         "//runtime/src/iree/testing:gtest_main",
@@ -706,12 +664,12 @@ iree_runtime_cc_test(
         ":pm4_dispatch_test_kernels",
         ":pm4_program",
         ":queue_primitives",
-        ":target_id",
         ":topology",
         ":vmem",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base/threading",
         "//runtime/src/iree/hal/drivers/amdgpu/abi",
+        "//runtime/src/iree/hal/executable/amdgpu:target_id",
         "//runtime/src/iree/testing:gtest",
         "//runtime/src/iree/testing:gtest_main",
     ],
@@ -752,7 +710,6 @@ iree_runtime_cc_benchmark(
     ],
     deps = [
         ":benchmark_flags",
-        ":executable_target",
         ":testdata_amdgpu_queue_benchmark",
         "//runtime/src/iree/async",
         "//runtime/src/iree/async/util:proactor_pool",
@@ -768,6 +725,7 @@ iree_runtime_cc_benchmark(
         "//runtime/src/iree/hal/drivers/amdgpu/device:dispatch",
         "//runtime/src/iree/hal/drivers/amdgpu/device:headers",
         "//runtime/src/iree/hal/drivers/amdgpu/registration",
+        "//runtime/src/iree/hal/executable/amdgpu:executable_target",
         "//runtime/src/iree/hal/memory:tlsf_pool",
         "//runtime/src/iree/io:file_handle",
     ],
@@ -853,10 +811,10 @@ iree_runtime_cc_test(
         ":libhsa",
         ":packet",
         ":queue_primitives",
-        ":target_id",
         ":topology",
         "//runtime/src/iree/base",
         "//runtime/src/iree/hal/drivers/amdgpu/abi",
+        "//runtime/src/iree/hal/executable/amdgpu:target_id",
         "//runtime/src/iree/testing:gtest",
         "//runtime/src/iree/testing:gtest_main",
     ],
