+++ runtime/src/iree/hal/drivers/amdgpu/util/CMakeLists.txt
@@ -305,57 +305,6 @@ endif()
 if(IREE_HAL_DRIVER_AMDGPU)
 iree_cc_library(
   NAME
-    target_id
-  HDRS
-    "target_id.h"
-  TEXTUAL_HDRS
-    "target_id_map.inl"
-  SRCS
-    "target_id.c"
-  DEPS
-    iree::base
-  PUBLIC
-)
-endif()
-
-if(IREE_HAL_DRIVER_AMDGPU)
-iree_cc_library(
-  NAME
-    executable_target
-  HDRS
-    "executable_target.h"
-  SRCS
-    "executable_target.c"
-  DEPS
-    ::target_id
-    iree::hal
-    iree::hal::utils::device_spec_builder
-  PUBLIC
-)
-endif()
-
-if(IREE_HAL_DRIVER_AMDGPU)
-iree_cc_test(
-  NAME
-    executable_target_test
-  SRCS
-    "executable_target_test.cc"
-  DEPS
-    ::executable_target
-    iree::testing::gtest
-    iree::testing::gtest_main
-  LABELS
-    "iree-build-requirement=runtime.hal.amdgpu"
-    "iree-run-requirement=runtime.resource.amd_gpu"
-    "runtime-resource=amd-gpu"
-  RESOURCE_GROUP
-    iree-hal-drivers-amdgpu-tests
-)
-endif()
-
-if(IREE_HAL_DRIVER_AMDGPU)
-iree_cc_library(
-  NAME
     code_object_target
   HDRS
     "code_object_target.h"
@@ -364,9 +313,9 @@ iree_cc_library(
   SRCS
     "code_object_target.c"
   DEPS
-    ::target_id
     iree::base
     iree::base::core_headers
+    iree::hal::executable::amdgpu::target_id
     iree::hal::utils::elf_format
   PUBLIC
 )
@@ -380,9 +329,9 @@ iree_cc_test(
     "code_object_target_test.cc"
   DEPS
     ::code_object_target
-    ::target_id
     iree::base
     iree::base::core_headers
+    iree::hal::executable::amdgpu::target_id
     iree::testing::gtest
     iree::testing::gtest_main
   LABELS
@@ -449,9 +398,9 @@ iree_cc_library(
     "pm4_dispatch.c"
   DEPS
     ::packet
-    ::target_id
     iree::base
     iree::hal::drivers::amdgpu::abi
+    iree::hal::executable::amdgpu::target_id
   PUBLIC
 )
 endif()
@@ -603,8 +552,8 @@ iree_cc_library(
   SRCS
     "device_library_target.c"
   DEPS
-    ::target_id
     iree::base
+    iree::hal::executable::amdgpu::target_id
   PUBLIC
 )
 endif()
@@ -627,8 +576,8 @@ iree_cc_library(
     ::pm4_dispatch
     ::pm4_program
     ::queue_primitives
-    ::target_id
     ::vmem
+    iree::hal::executable::amdgpu::target_id
   PUBLIC
 )
 endif()
@@ -746,26 +695,6 @@ iree_cc_test(
     "device_library_target_test.cc"
   DEPS
     ::device_library_target
-    iree::base
-    iree::testing::gtest
-    iree::testing::gtest_main
-  LABELS
-    "iree-build-requirement=runtime.hal.amdgpu"
-    "iree-run-requirement=runtime.resource.amd_gpu"
-    "runtime-resource=amd-gpu"
-  RESOURCE_GROUP
-    iree-hal-drivers-amdgpu-tests
-)
-endif()
-
-if(IREE_HAL_DRIVER_AMDGPU)
-iree_cc_test(
-  NAME
-    target_id_test
-  SRCS
-    "target_id_test.cc"
-  DEPS
-    ::target_id
     iree::base
     iree::testing::gtest
     iree::testing::gtest_main
@@ -1042,12 +971,12 @@ iree_cc_test(
     ::pm4_dispatch_test_kernels
     ::pm4_program
     ::queue_primitives
-    ::target_id
     ::topology
     ::vmem
     iree::base
     iree::base::threading
     iree::hal::drivers::amdgpu::abi
+    iree::hal::executable::amdgpu::target_id
     iree::testing::gtest
     iree::testing::gtest_main
   LABELS
@@ -1115,7 +1044,6 @@ iree_cc_binary_benchmark(
     "queue_benchmark.cc"
   DEPS
     ::benchmark_flags
-    ::executable_target
     ::testdata_amdgpu_queue_benchmark
     iree::async
     iree::async::util::proactor_pool
@@ -1131,6 +1059,7 @@ iree_cc_binary_benchmark(
     iree::hal::drivers::amdgpu::headers
     iree::hal::drivers::amdgpu::queue_affinity
     iree::hal::drivers::amdgpu::registration
+    iree::hal::executable::amdgpu::executable_target
     iree::hal::memory::tlsf_pool
     iree::io::file_handle
     iree::third_party::google_benchmark
@@ -1274,10 +1203,10 @@ iree_cc_test(
     ::libhsa
     ::packet
     ::queue_primitives
-    ::target_id
     ::topology
     iree::base
     iree::hal::drivers::amdgpu::abi
+    iree::hal::executable::amdgpu::target_id
     iree::testing::gtest
     iree::testing::gtest_main
   LABELS
