+++ runtime/src/iree/hal/drivers/amdgpu/util/code_object_target.h
@@ -7,7 +7,7 @@
 #ifndef IREE_HAL_DRIVERS_AMDGPU_UTIL_CODE_OBJECT_TARGET_H_
 #define IREE_HAL_DRIVERS_AMDGPU_UTIL_CODE_OBJECT_TARGET_H_
 
-#include "iree/hal/drivers/amdgpu/util/target_id.h"
+#include "iree/hal/executable/amdgpu/target_id.h"
 
 #ifdef __cplusplus
 extern "C" {
