+++ runtime/src/iree/hal/drivers/amdgpu/util/pm4_dispatch_live_test.cc
@@ -22,9 +22,9 @@
 #include "iree/hal/drivers/amdgpu/util/pm4_dispatch.h"
 #include "iree/hal/drivers/amdgpu/util/pm4_dispatch_test_kernels.h"
 #include "iree/hal/drivers/amdgpu/util/pm4_program.h"
-#include "iree/hal/drivers/amdgpu/util/target_id.h"
 #include "iree/hal/drivers/amdgpu/util/topology.h"
 #include "iree/hal/drivers/amdgpu/util/vmem.h"
+#include "iree/hal/executable/amdgpu/target_id.h"
 #include "iree/testing/gtest.h"
 #include "iree/testing/status_matchers.h"
