+++ runtime/src/iree/hal/drivers/hip/BUILD.bazel
@@ -112,7 +112,7 @@ iree_runtime_cc_library(
         "//runtime/src/iree/base",
         "//runtime/src/iree/base:core_headers",
         "//runtime/src/iree/hal",
-        "//runtime/src/iree/hal/drivers/amdgpu/util:executable_target",
+        "//runtime/src/iree/hal/executable/amdgpu:executable_target",
         "//runtime/src/iree/hal/utils:device_spec_builder",
     ],
 )
