+++ runtime/src/iree/hal/drivers/hip/CMakeLists.txt
@@ -95,7 +95,7 @@ iree_cc_library(
     iree::base
     iree::base::core_headers
     iree::hal
-    iree::hal::drivers::amdgpu::util::executable_target
+    iree::hal::executable::amdgpu::executable_target
     iree::hal::utils::device_spec_builder
   PUBLIC
 )
