+++ runtime/src/iree/hal/executable/CMakeLists.txt
@@ -1,0 +1,12 @@
+################################################################################
+# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
+# runtime/src/iree/hal/executable/BUILD.bazel                                  #
+#                                                                              #
+# Add CMake-only content below the preserve marker at the end of this file.    #
+#                                                                              #
+# To disable autogeneration for this file entirely, delete this header.        #
+################################################################################
+
+iree_add_all_subdirs()
+
+### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
