+++ runtime/src/iree/hal/executable/amdgpu/CMakeLists.txt
@@ -1,0 +1,63 @@
+################################################################################
+# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
+# runtime/src/iree/hal/executable/amdgpu/BUILD.bazel                           #
+#                                                                              #
+# Add CMake-only content below the preserve marker at the end of this file.    #
+#                                                                              #
+# To disable autogeneration for this file entirely, delete this header.        #
+################################################################################
+
+iree_add_all_subdirs()
+
+iree_cc_library(
+  NAME
+    target_id
+  HDRS
+    "target_id.h"
+  TEXTUAL_HDRS
+    "target_id_map.inl"
+  SRCS
+    "target_id.c"
+  DEPS
+    iree::base
+  PUBLIC
+)
+
+iree_cc_test(
+  NAME
+    target_id_test
+  SRCS
+    "target_id_test.cc"
+  DEPS
+    ::target_id
+    iree::base
+    iree::testing::gtest
+    iree::testing::gtest_main
+)
+
+iree_cc_library(
+  NAME
+    executable_target
+  HDRS
+    "executable_target.h"
+  SRCS
+    "executable_target.c"
+  DEPS
+    ::target_id
+    iree::hal
+    iree::hal::utils::device_spec_builder
+  PUBLIC
+)
+
+iree_cc_test(
+  NAME
+    executable_target_test
+  SRCS
+    "executable_target_test.cc"
+  DEPS
+    ::executable_target
+    iree::testing::gtest
+    iree::testing::gtest_main
+)
+
+### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
