+++ runtime/src/iree/hal/executable/amdgpu/executable_target_test.cc
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/hal/drivers/amdgpu/util/executable_target.h"
+#include "iree/hal/executable/amdgpu/executable_target.h"
 
 #include "iree/testing/gtest.h"
 #include "iree/testing/status_matchers.h"
