+++ runtime/src/iree/hal/executable/amdgpu/target_id.h
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef IREE_HAL_DRIVERS_AMDGPU_UTIL_TARGET_ID_H_
-#define IREE_HAL_DRIVERS_AMDGPU_UTIL_TARGET_ID_H_
+#ifndef IREE_HAL_EXECUTABLE_AMDGPU_TARGET_ID_H_
+#define IREE_HAL_EXECUTABLE_AMDGPU_TARGET_ID_H_
 
 #include "iree/base/api.h"
 
@@ -172,4 +172,4 @@ iree_status_t iree_hal_amdgpu_target_compatibility_format(
 }  // extern "C"
 #endif  // __cplusplus
 
-#endif  // IREE_HAL_DRIVERS_AMDGPU_UTIL_TARGET_ID_H_
+#endif  // IREE_HAL_EXECUTABLE_AMDGPU_TARGET_ID_H_
