+++ runtime/src/iree/hal/executable/amdgpu/target_id_map.inl
@@ -1,6 +1,6 @@
 // Generated by build_tools/amdgpu/target_map.py.
 // Do not edit directly; edit the Python source tables and regenerate.
-// Output: runtime/src/iree/hal/drivers/amdgpu/util/target_id_map.inl
+// Output: runtime/src/iree/hal/executable/amdgpu/target_id_map.inl
 //
 // Included inside iree_hal_amdgpu_target_id_mappings.
